refactor: reorgnize test cases
